feature/139 - Move position model conversion functions to individual external helpers
