+++ src/assets/scripts/client/base/dynamicPositionToStaticPositionHelper.js
@@ -1,0 +1,16 @@
+import StaticPositionModel from './StaticPositionModel';
+
+/**
+ * Accepts a `DynamicPositionModel` and returns a `StaticPositionModel` with the same location
+ *
+ * @function convertDynamicPositionToStatic
+ * @param dynamicPositionModel {DynamicPositionModel}
+ * @return {StaticPositionModel}
+ */
+export const convertDynamicPositionToStatic = (dynamicPositionModel) => {
+    const staticPositionModel = new StaticPositionModel(dynamicPositionModel.gps,
+        dynamicPositionModel.referencePostion, dynamicPositionModel.magnetic_north
+    );
+
+    return staticPositionModel;
+};
